backend/app/requirements.txt
# Core dependencies for model training and inference
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
albumentations==2.0.5
pandas==2.2.3
timm==1.0.15
scikit-learn==1.6.1
pillow==10.2.0

# For face detection (YOLOv11 and MTCNN)
ultralytics==8.3.0  # For YOLOv11
facenet-pytorch==2.6.0  # For MTCNN
opencv-python==4.11.0.86

# For web app deployment (FastAPI)
fastapi==0.115.2
uvicorn==0.32.0  # For running FastAPI
python-multipart==0.0.12  # For handling file uploads in FastAPI

# Additional utilities (if used in data generation)
huggingface-hub==0.29.1  # For accessing Stable Diffusion models
diffusers==0.32.2  # For Stable Diffusion
transformers==4.49.0  # For Hugging Face models
safetensors==0.5.3  # For loading model weights safely